UI improvements for the form configuration page
- Split the form configuration UI into tabs for XML and JSON
- Use vertical forms
- Use angular ng-click instead of jquery events
- Clean up code to use latest best practices

medic/medic-webapp#3559
