+++ templates/partials/configuration_forms.html
@@ -1,92 +1,9 @@
-<div class="col-sm-12">
-  <form id="forms-upload-xform">
-    <div>
-      <fieldset>
-        <legend translate>upload.xform.title</legend>
-        <div class="alert alert-danger alert-dismissible" role="alert" ng-show="xform.error">
-          <button type="button" class="close" aria-label="Close" ng-click="xform.error = false"><span aria-hidden="true">&times;</span></button>
-          <span translate>Upload failed</span>
-        </div>
-        <div class="alert alert-success" role="alert" ng-show="xform.success" translate>Upload succeeded</div>
-        <label><span translate>upload.xform.xml</span>
-          <input class="form uploader" type="file" accept="application/xml"/>
-        </label>
-        <label><span translate>upload.xform.context</span>
-          <input class="meta uploader" type="file" accept="application/json"/>
-        </label>
-        <button type="button" class="btn btn-default upload" ng-disabled="xform.uploading">
-          <i class="fa fa-arrow-up"></i>
-          <span translate>Import</span>
-        </button>
-        <div class="loader" ng-show="xform.uploading"></div>
-        <span class="help-block" translate>upload.xform.help</span>
-        <div class="hide">
-        </div>
-      </fieldset>
-    </div>
-  </form>
-
-  <div class="section">
-    <fieldset>
-      <legend translate>Installed Forms</legend>
-      <ul>
-        <li ng-repeat="form in xForms" class="row">
-          <div class="col-sm-6">{{form._id}}</div>
-          <div class="col-sm-6">{{form.title | translateFrom}}</div>
-        </li>
-      </ul>
-    </fieldset>
-  </div>
-
-  <form id="forms-upload-json">
-    <div>
-      <fieldset>
-        <legend translate>upload.json.forms.title</legend>
-        <div class="alert alert-danger alert-dismissible" role="alert" ng-show="json.error">
-          <button type="button" class="close" aria-label="Close" ng-click="json.error = false"><span aria-hidden="true">&times;</span></button>
-          <span translate>Upload failed</span>
-        </div>
-        <div class="alert alert-success" role="alert" ng-show="json.success" translate>Upload succeeded</div>
-        <button type="button" class="btn btn-default choose" ng-disabled="json.uploading">
-          <i class="fa fa-arrow-up"></i>
-          <span translate>Import</span>
-        </button>
-        <div class="loader" ng-show="json.uploading"></div>
-        <span class="help-block" translate>upload.json.forms.help</span>
-        <div class="hide">
-          <input class="uploader" type="file" accept="application/json">
-        </div>
-      </fieldset>
-    </div>
-  </form>
-
-  <div class="section">
-    <fieldset>
-      <legend translate>Installed Forms</legend>
-      <ul>
-        <li ng-repeat="form in forms" class="row">
-          <div class="col-sm-1" ng-bind-html="form.meta.icon | resourceIcon"></div>
-          <div class="col-sm-1">{{form.meta.code}}</div>
-          <div class="col-sm-10">{{form.meta.label | translateFrom}}</div>
-        </li>
-      </ul>
-      <a class="btn btn-default" ng-attr-download="{{download.name}}" ng-href="{{download.url}}">
-        <i class="fa fa-arrow-down"></i>
-        <span translate>Download</span>
-      </a>
-    </fieldset>
-  </div>
-
-  <form class="section">
-    <div>
-      <fieldset>
-        <legend translate>reporter.section.title</legend>
-        <a class="btn btn-default" href="/medic-reporter/_design/medic-reporter/_rewrite/">
-          <i class="fa fa-external-link"></i>
-          <span translate>reporter.link</span>
-        </a>
-        <span class="help-block" translate>reporter.section.help</span>
-      </fieldset>
-    </div>
-  </form>
-</div>
+<ul class="nav nav-tabs">
+  <li role="presentation" ui-sref-active="active">
+    <a ui-sref="configuration.forms.xml" translate>forms.xml</a>
+  </li>
+  <li role="presentation" ui-sref-active="active">
+    <a ui-sref="configuration.forms.json" translate>forms.json</a>
+  </li>
+</ul>
+<div class="tab-content" ui-view="tab"></div>
