+++ templates/partials/configuration_forms_xml.html
@@ -1,0 +1,37 @@
+<form id="forms-upload-xform">
+  <legend translate>Upload Forms</legend>
+
+  <div class="form-group">
+    <label translate>upload.xform.xml</label>
+    <input class="form uploader" type="file" accept="application/xml"/>
+    <span class="help-block" translate>upload.xform.xml.help</span>
+  </div>
+
+  <div class="form-group">
+    <label translate>upload.xform.context</label>
+    <input class="meta uploader" type="file" accept="application/json"/>
+    <span class="help-block" translate>upload.xform.meta.help</span>
+  </div>
+
+  <div class="form-actions">
+    <a class="btn btn-primary upload" ng-disabled="status.uploading" ng-click="upload()">
+      <i class="fa fa-arrow-up"></i>
+      <span translate>Import</span>
+    </a>
+    <div class="error" ng-show="status.error" translate>Upload failed</div>
+    <div class="loader" ng-show="status.uploading"></div>
+  </div>
+</form>
+
+<div class="section">
+  <fieldset>
+    <legend translate>Installed Forms</legend>
+    <ul>
+      <li ng-repeat="form in xForms | orderBy:'_id' track by form._id" class="row">
+        <div class="col-sm-1" ng-bind-html="form.icon | resourceIcon"></div>
+        <div class="col-sm-6">{{form._id}}</div>
+        <div class="col-sm-5">{{form.title | translateFrom}}</div>
+      </li>
+    </ul>
+  </fieldset>
+</div>
