+++ translations/messages-en.properties
@@ -901,3 +901,7 @@ upload.xform.help = The files should contain an XForm and it's associated JSON c
 Upload\ Forms = Upload forms
 Upload\ forms\ help = The file should contain a JSON encoded list of form definitions. Any existing forms will be overwritten.
 debug.mode = Debug
+forms.xml = XML
+forms.json = JSON
+upload.xform.xml.help = The xform definition in XML.
+upload.xform.meta.help = The meta file in JSON format including the icon, title, context, and so on.
